backend/requirements.txt
scipy==1.13.1
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.116.1
Flask==3.1.2
flask-cors==6.0.1
ftfy==6.3.1
gensim==4.3.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
MarkupSafe==3.0.2
numpy==1.26.4
openai==1.102.0
opencv-python-headless==4.11.0.86 
outcome==1.3.0.post0
pdf2image==1.17.0
pillow==11.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pypdf==6.0.0
PyPDF2==3.0.1
PySocks==1.7.1
pythainlp==5.0.5
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.14.0
requests==2.32.5
scikit-learn==1.3.2
selenium==4.35.0
typhoon-ocr==0.3.8
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
undetected-chromedriver
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0










